% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identifiable.R
\name{causal_effect}
\alias{causal_effect}
\title{Causal Effect Identification}
\usage{
causal_effect(y, x = character(0), z = character(0), v = integer(0), g)
}
\arguments{
\item{y}{A \code{character} vector of outcome variables.}

\item{x}{A \code{character} vector of intervention variables.}

\item{z}{An optional \code{character} vector of conditioning variables}

\item{v}{An optional named \code{integer} vector giving the value assignments
for observed variables in the model.}

\item{g}{A \code{dag} object depicting the causal model.}
}
\description{
Identify a causal effect of the form \eqn{p(y|do(x),z)}.
}
