% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{export_graph}
\alias{export_graph}
\title{Export Graph}
\usage{
export_graph(
  g,
  type = c("dagitty", "causaleffect", "dosearch"),
  use_bidirected = TRUE,
  ...
)
}
\arguments{
\item{g}{An object of class \code{DAG}.}

\item{type}{A character string matching one of the following:
"dagitty", "causaleffect" or "dosearch".}

\item{...}{Additional arguments passed to \code{format}
for formatting vertex labels.}
}
\description{
Convert a valid graph object to a supported external format.
}
