% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identifiable.R
\name{query}
\alias{query}
\alias{print.query}
\title{Query Objects}
\usage{
\method{print}{query}(x, ...)
}
\arguments{
\item{x}{A \code{query} object}

\item{...}{Not used.}
}
\description{
Objects of class \code{query} describe the output of \code{identifiable} and
\code{causal_effect}. They are \code{list} objects with a custom \code{print} method and
contain data related to the identifiability results. See
\link{identifiable} and \link{causal_effect} for details.
}
