% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractors.R
\name{trivial_conflict}
\alias{trivial_conflict}
\title{Determines whether a counterfactual variable conflicts with a
counterfactual conjunction}
\usage{
trivial_conflict(y, gamma)
}
\arguments{
\item{y}{A \code{counterfactual_variable} object.}

\item{gamma}{A \code{Counterfactual_conjunction} object.}
}
\value{
A \code{list} of conflicting variables.
}
\description{
Determines whether a counterfactual variable conflicts with a
counterfactual conjunction
}
