% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probability.R
\name{format.probability}
\alias{format.probability}
\title{Format a \code{probability} object representing \eqn{p(y|do(x))}.}
\usage{
\method{format}{probability}(x, use_primes = TRUE, use_do = FALSE, ...)
}
\arguments{
\item{x}{A \code{probability} object.}

\item{use_primes}{A \code{logical} value. If \code{TRUE} (the default), any value
assignment of a counterfactual variable with \code{obs} will be formatted with
as many primes in the superscript as the value of \code{obs}, e.g.,
\code{obs = 0} outputs \code{"y"}, \code{obs = 1} outputs \code{"y'"},
\code{obs = 2} outputs \code{"y''"} and so forth. The alternative when \code{FALSE} is
to simply denote the \code{obs} value via superscript directly as
\code{"y^{(obs)}"}, where obs is evaluated.}

\item{use_do}{A \code{logical} value. If \code{TRUE}, the explicit do-operation is used
to denote interventional probabilities (e.g., \eqn{p(y|do(x))}).
If \code{FALSE} (the default), the subscript notation is used instead
(e.g., \eqn{p_x(y})).}

\item{...}{Additional arguments passed to \code{format}.}
}
\value{
An character representation of the \code{probability} object
in LaTeX syntax.
}
\description{
Format a \code{probability} object representing \eqn{p(y|do(x))}.
}
\examples{
# Example from Makhlouf, Zhioua and Palamidessi (2021)
g2 <- dag("C -> A -> Y; C -> Y")
v1 <- cf("Y", 0, c(A = 1))
v2 <- cf("A", 0)
c1 <- conj(v1)
c2 <- conj(v2)
p <- identifiable(g2, c1, c2)$prob

# Default, using primes and subscript notation
format(p)

# Without primes, no do-operator
format(p, use_primes = FALSE)

# Primes, with do-operator
format(p, use_do = TRUE)

# Without primes, with do-operator
format(p, use_primes = FALSE, use_do = TRUE)

}
\references{
Makhlouf, K., Zhioua, S. and Palamidessi, C. (2021).
Survey on causal-based machine learning fairness notions.
\emph{arXiv:2010.09553}
}
