% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functional.R
\name{functional}
\alias{functional}
\title{Identifying Functional of a Counterfactual Query}
\usage{
functional(
  sumset = NULL,
  summand = NULL,
  terms = NULL,
  numerator = NULL,
  denominator = NULL
)
}
\arguments{
\item{sumset}{A \code{list} of objects of class \code{counterfactual_variable}
(without interventions and with value assignments).
If the probability depicts marginalization, \code{sumset}
defines the set of variables to be marginalized over.}

\item{summand}{A \code{functional} or a \code{probability} object. If \code{sumset}
is not \code{NULL}, this defines the probability being marginalized.}

\item{terms}{A \code{list} of \code{functional} objects if the object in question
is meant to represent a product of terms.}

\item{numerator}{A \code{functional} or a \code{probability} object.
If the functional represents a conditional probability that
cannot be expressed simply in terms of the set of inputs \eqn{P*},
this is the numerator of the quotient representation.}

\item{denominator}{A \code{functional} or a \code{probability} object.
The denominator of the quotient representation.}
}
\value{
An object of class \code{functional}, which is a \code{list} containing
all of the arguments of the constructor.
}
\description{
Identifying functionals are more complicated probabilistic expressions
that cannot be expressed as simple observational or interventional
}
\details{
When formatted via \code{print} or \code{format}, the arguments are
prioritized in the following order if conflicting definitions are given:
(\code{sumset}, \code{summand}), \code{terms}, (\code{numerator}, \code{denominator}).
}
