% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identifiable.R
\name{identifiable}
\alias{identifiable}
\title{Identify a Counterfactual Query}
\usage{
identifiable(g, gamma, delta = NULL)
}
\arguments{
\item{g}{A \code{DAG} object describing the causal graph
(or an object that can be coerced, see \code{\link[=import_graph]{import_graph()}}.}

\item{gamma}{An object of class \code{CounterfactualConjunction}
representing the counterfactual causal query.}

\item{delta}{An object of class \code{CounterfactualConjunction}
representing the conditioning conjunction (optional).}
}
\value{
A list containing one or more of the following:
\itemize{
\item \code{id} A logical value that is \code{TRUE} if the query is identifiable and
\code{FALSE} otherwise. Note that in cases where \code{gamma} is itself
inconsistent, the query will be identifiable, but with probability 0.
\item \code{prob} An object of class \code{Probability} giving the formula of the query in
LaTeX syntax via format or print, if identifiable.
This expression is given in terms of \eqn{P*},
the set of all interventional distributions over \code{g}. For tautological
statements, the resulting probability is 1, and for inconsistent
statements, the resulting probability is 0.
\item \code{undefined} A logical value that is \code{TRUE} if
a conditional conjunction \eqn{p(\gamma|\delta)} is undefined,
for example when \eqn{p(\delta) = 0}, and \code{FALSE} otherwise.
}
}
\description{
Determine the identifiability of a (conditional) counterfactual conjunction.
}
\details{
To identify a non-conditional conjunction \eqn{p(\gamma)}, the argument
\code{delta} should be \code{NULL}.

To identify a conditional conjunction \eqn{p(\gamma|\delta)}, both \code{gamma}
and \code{delta} should be specified.

First, a parallel worlds graph is constructed based on the query. In a
parallel worlds graph, for each \eqn{do}-action that appears in \eqn{gamma}
(and \eqn{delta}) a copy of the original graph is created with the new
observational variables attaining their post-interventional values
but sharing the latent variables. From this graph, a counterfactual graph
is derived, where each variable is unique, which might not be the case
in a parallel worlds graph.

Finally, the ID* (or IDC*) algorithm is applied to determine identifiability
of the query. Similar to the ID and IDC algorithms for causal effects,
these algorithms exploit the so called c-component factorization to split
the query into smaller subproblems, which are then solved recursively.
}
\examples{
# Examples that appears in Shpitser and Pearl (2008)
g1 <- dag("X -> W -> Y <- Z <- D X <-> Y")
g2 <- dag("X -> W -> Y <- Z <- D X <-> Y X -> Y")
v1 <- cf("Y", 0, c(X = 0))
v2 <- cf("X", 1)
v3 <- cf("Z", 0, c(D = 0))
v4 <- cf("D", 0)
c1 <- conj(v1)
c2 <- conj(v2, v3, v4)
c3 <- conj(v1, v2, v3, v4)

# Identifiable conditional conjunction
identifiable(g1, c1, c2)

# Identifiable conjunction
identifiable(g1, c3)

# Non-identifiable conjunction
identifiable(g2, c3)
}
\references{
Shpister, I. and Pearl, J. (2006). Identification of joint interventional
distributions in semi-Markovian causal models.
In \emph{21st National Conference on Artificial Intelligence}

Shpitser, I. and Pearl, J. (2008). Complete identification
methods for the causal hierarchy. \emph{Journal of Machine Learning Research},
\strong{9(64)}:1941--1979.
}
\seealso{
\code{\link[=dag]{dag()}}, \code{\link[=CounterfactualVariable]{CounterfactualVariable()}},
\code{\link[=CounterfactualConjunction]{CounterfactualConjunction()}}
}
