% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{import_graph}
\alias{import_graph}
\title{Import Graph}
\usage{
import_graph(x)
}
\arguments{
\item{x}{A graph object in a valid external format, see details.}
}
\value{
A \code{DAG} object if successful.
}
\description{
Import and construct a valid DAG from an external format.
}
\details{
Argument \code{x} accepts \code{\link[dagitty:dagitty]{dagitty::dagitty()}} graphs, igraph-graphs
in the \code{\link[causaleffect:causaleffect-package]{causaleffect::causaleffect()}} package syntax and character strings
in the \code{\link[dosearch:dosearch]{dosearch::dosearch()}} package syntax.
}
