% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probability.R
\name{probability}
\alias{probability}
\title{Symbolic Probability Distributions}
\usage{
probability(val = NULL, var = NULL, do = NULL, cond = NULL)
}
\arguments{
\item{val}{An \code{integer} value of either 0 or 1 for almost sure events.}

\item{var}{A \code{list} of objects of class \code{counterfactual_variable}
(without interventions and with value assignments).
\code{var} defines the observations \eqn{y} in \eqn{p(y|do(x),z)}.}

\item{do}{A \code{list} of \code{counterfactual_variable} variable objects
(without interventions and with value assignments). Defines the
intervention set \eqn{x} in \eqn{p(y|do(x),z)}.}

\item{cond}{A \code{list} of \code{counterfactual_variable} variable objects
(without interventions and with value assignments).Defines the
conditioning set \eqn{z} in \eqn{p(y|do(x),z)}.}
}
\value{
An object of class \code{probability}, which is a \code{list} containing
all of the arguments of the constructor.
}
\description{
Defines an interventional or observational (conditional) probability
\eqn{p(y|do(x),z)}. For formatting options, see \code{\link[=format.probability]{format.probability()}}.
}
\seealso{
\code{\link[=counterfactual_variable]{counterfactual_variable()}}, \code{\link[=functional]{functional()}}
}
